requirements.txt
aniso8601==9.0.1
astroid==2.6.2
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.2.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.2
mccabe==0.6.1
mongoengine==0.23.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.3
pymongo==3.11.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
